+++ Assets/RevCore/Foundation/Samples~/ServiceLocatorSample/IAudioService.cs
@@ -1,0 +1,8 @@
+namespace RevCore.Foundation.Samples
+{
+    public interface IAudioService
+    {
+        void PlaySfx(string clipName);
+        void StopAll();
+    }
+}